ml/train/requirements.txt
# ML Training Requirements
# Install: pip install -r ml/train/requirements.txt

torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
datasets>=2.14.0
wandb>=0.16.0
tensorboard>=2.14.0
trl>=0.7.0
scipy>=1.11.0
sentencepiece>=0.1.99
protobuf>=4.24.0
safetensors>=0.4.0
pyyaml>=6.0
tqdm>=4.66.0
